Rename duplicate scone layer files -> _bis
