+++ topomodelx/nn/simplicial/scone_layer_bis.py
